requirements.txt
aiogram==3.5.0
pydantic-settings==2.3.1
pydantic==2.7.4
python-dotenv==1.0.1
asyncpg==0.29.0
aiohttp==3.9.5
